requirements.txt
gym
jedi==0.17.2
matplotlib
nbdime
numpy
opencv-python
pandas
pytorch-lightning-bolts
pytest
pytest-cov
pdbpp
rasterio
scikit-learn
torch
torchvision
wandb
xarray
geopandas
tqdm
